# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API V1",
        "version": "0.1"
    },
    "paths": {
...
        "api/v1/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Retrive a listing of the Transactions",
                "operationId": "c97a363fb6209c4eb89bf72067a8c46b",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Store new Transaction",
                "operationId": "5bc125dd97914b35e0777ac87c71fb7d",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/v1/transactions/{uuid}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get an existing Transaction",
                "operationId": "e8065b04ab2d9710fa378d57d0601cff",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update an existing Transaction",
                "operationId": "0569b630423d4438e6e0b441feec6385",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete existing Transaction",
                "operationId": "03ff35daae36479834590d9ffecba961",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreTransactionRequest": {
                "title": "StoreTransactionRequest",
                "properties": {
                    "amount": {
                        "type": "string",
                        "example": "example"
                    },
                    "description": {
                        "type": "string",
                        "example": "example"
                    },
                    "currency": {
                        "type": "string",
                        "example": "example"
                    },
                    "status": {
                        "type": "string",
                        "example": "example"
                    },
                    "issuer": {
                        "type": "string",
                        "example": "example"
                    },
                    "transaction_date": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
            "UpdateTransactionRequest": {
                "title": "UpdateTransactionRequest",
                "properties": {
                    "description": {
                        "type": "string",
                        "example": "example"
                    },
                    "status": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "title": "Transaction",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                    },
                    "amount": {
                        "type": "string",
                        "example": "example"
                    },
                    "description": {
                        "type": "string",
                        "example": "example"
                    },
                    "currency": {
                        "type": "string",
                        "example": "example"
                    },
                    "status": {
                        "type": "string",
                        "example": "example"
                    },
                    "issuer": {
                        "type": "string",
                        "example": "example"
                    },
                    "transaction_date": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            }
        }
    }
}